strace.c: use snprintf instead of sprintf
We can never be cautious enough.

* strace.c (strerror, newoutf, print_debug_info): Replace sprintf with
snprintf, providing the size of the respective output buffer.
(attach_tcb): Add task_path variable with the output format. Declare
procdir as static const, use its size as a size argument in the
replacing snprintf call, as well as task_path for the format string.