Improve handling of unexpected tracees
When receiving a ptrace stop from an unexpected child, handle it in the most
transparent way possible:
 - Try to detach it instead of CONT'ing it.
 - Send it the signal with which it has been stopped.
This should hopefully help to deal with processes which has been created with
misused CLONE_PTRACE flag set.

* strace.c (maybe_allocate_tcb) <WIFSTOPPED(status) && !followfork>: Calculate
signal similarly to the way next_event does it, forward it to the unexpected
tracee, and detach it.