xlat/gen.sh: speedup xlat generation
This commit does three things:
 - Increases the number of concurrent jobs to twice the CPU count
 - Creates circular buffer, so instead of running multiple jobs at once,
   we try to always have about the same number of jobs being run
   concurrently.
 - Runs gen_git and gen_make concurrently in order to squeeze in
   one more bit of concurrency.

With my deeply scientific measurements, this approach gives up to 15%
speedup with dash and about 30-40% with bash as /bin/sh on 4-core
machine.

* xlat/gen.sh (main): Declare pids local variable, append pid of every
run job to it; increase the limit of jobs to ncpus * 2; when the limit
is reached, wait for the first pid in pids instead of resetting jobs
to zero and waiting for all the run jobs; put gen_git and gen_make
into background.